# Conflicts:
#	src/main/java/com/ssaw/InventoryManagement/entity/SecuritiesClosedPayInventory.java

+++ src/main/java/com/ssaw/InventoryManagement/entity/SecuritiesClosedPayInventory.java
@@ -1,7 +1,5 @@
 package com.ssaw.InventoryManagement.entity;
 
-import java.math.BigDecimal;
-
 /**
  * ClassName:    SecuritiesClosedPayInventory
  * Package:    com.ssaw.InventoryManagement.entity
@@ -26,7 +24,7 @@ public class SecuritiesClosedPayInventory {
     //业务状态 1流入，-1流出
     private int flag;
     //总金额*
-    private BigDecimal totalPrice;
+    private double totalPrice;
     //备注
     private String securitiesClosedPayDesc;
     //期初标志 是否从其他系统导入得期初数据 0：不是 1：是
@@ -35,6 +33,7 @@ public class SecuritiesClosedPayInventory {
     public SecuritiesClosedPayInventory(){
 
     }
+
 
     /**
      *
@@ -48,8 +47,7 @@ public class SecuritiesClosedPayInventory {
      * @param securitiesClosedPayDesc   备注
      * @param securityPeriodFlag    期初标志 是否从其他系统导入得期初数据 0：不是 1：是
      */
-
-    public SecuritiesClosedPayInventory(String securitiesClosedPayInventoryId, String dateTime, String fundId, String securitiesId, int securitiesType, int flag, BigDecimal totalPrice, String securitiesClosedPayDesc, int securityPeriodFlag) {
+    public SecuritiesClosedPayInventory(String securitiesClosedPayInventoryId,String dateTime, String fundId, String securitiesId, int securitiesType, int flag, double totalPrice, String securitiesClosedPayDesc, int securityPeriodFlag) {
         this.securitiesClosedPayInventoryId = securitiesClosedPayInventoryId;
         this.dateTime = dateTime;
         this.fundId = fundId;
@@ -60,8 +58,6 @@ public class SecuritiesClosedPayInventory {
         this.securitiesClosedPayDesc = securitiesClosedPayDesc;
         this.securityPeriodFlag = securityPeriodFlag;
     }
-
-
 
     public String getSecuritiesClosedPayInventoryId() {
         return securitiesClosedPayInventoryId;
@@ -112,11 +108,11 @@ public class SecuritiesClosedPayInventory {
         this.flag = flag;
     }
 
-    public BigDecimal getTotalPrice() {
+    public double getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(BigDecimal totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
